refactor(workflow-executor): export TriggerType once from index, drop comments
Export TriggerType directly from its home (validated/execution) in the
barrel instead of re-exporting through execution-context, and remove the
explanatory comments on the two trigger-type enums.
