+++ packages/workflow-executor/src/index.ts
@@ -58,7 +58,7 @@ export type {
   ExecutionContext,
 } from './types/execution-context';
 
-export { TriggerType } from './types/execution-context';
+export { TriggerType } from './types/validated/execution';
 
 export type {
   AgentPort,
